chore: Rework code to get nightlies to discriminate between schedules
The only way to distinguish which "scheduled" pipelines are triggered
from which configured schedule, is to use the dedicated API endpoint
`GET /projects/:id/pipeline_schedules/:pipeline_schedule_id/pipelines`.

ref: https://docs.gitlab.com/ee/api/pipeline_schedules.html#get-all-pipelines-triggered-by-a-pipeline-schedule

Unfortunately, this endpoint does not seem to be implemented in GraphQL
API, nor there is a way to connect from project -> pipelineSchedules ->
pipelines, so we have to fallback to the REST API.

On top of it, this endpoint does not support keyset-based pagination, so
the code does a fair amount of API calls.

Originally we assumed that we could just keep the GraphQL query and then
filter based on each pipeline variables, but it turns out that the
variables endpoint only includes the user variables, so the
`NIGHTLY_RUN` variable injected from the schedule is nowhere to be seen.

Ticket: QA-539

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>