chore: Remove circular dependency across components hierarchy
Top level `nightlies.js` was defining the mysterious `order` array with
the sorted names of the pipelines to then be imported by `CalendarView`
resulting in a circular import (which seems like js handles just fine).

In any case closer inspection of the code revealed that the only use of
it was to guarantee a given order of the `Dot` components across days in
the calendar, which can be achieved by sorting before hand before
rendering the top level `Nightlies` comopnent.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>